Modularise popup.js and refactor utility functions into separate files
